+++ src/mappings/parachain-staking/index.ts
@@ -1,5 +1,6 @@
 import { HistoricalAccountBalance } from '../../model';
-import { _Asset, extrinsicFromEvent } from '../../helper';
+import { _Asset } from '../../consts';
+import { extrinsicFromEvent } from '../../helper';
 import { Block, Event } from '../../processor';
 import { decodeRewardedEvent } from './decode';
