+++ src/server-extension/resolvers/marketMetadata.ts
@@ -1,7 +1,7 @@
 import { Arg, Field, Int, ObjectType, Query, Resolver } from 'type-graphql';
 import type { EntityManager } from 'typeorm';
 import { Market } from '../../model/generated';
-import { CacheHint } from '../../helper';
+import { CacheHint } from '../../consts';
 import { Cache } from '../../util';
 import { marketMetadata } from '../query';
